+++ src/components/Link/LinkKakaotalkShare.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from "react";
 import { useLocation } from "react-router-dom";
 
-import { kakaoSDKKey } from "loadenv";
+import { kakaoSDKKey } from "tools/loadenv";
 
 type LinkKakaotalkShareProps = {
   children: React.ReactNode;
